updated pharmacy type to handle view model
