+++ src/server/HelloCodeaholics.Api/Controllers/ErrorController.cs
@@ -1,13 +1,11 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 
-namespace HelloCodeaholics.Api.Controllers
+namespace HelloCodeaholics.Api.Controllers;
+
+[ApiController]
+public class ErrorController : ControllerBase
 {
-    [ApiController]
-    public class ErrorController : ControllerBase
-    {
-        [ApiExplorerSettings(IgnoreApi = true)]
-        [Route("/error")]
-        public IActionResult HandleError() => Problem();
-    }
+    [ApiExplorerSettings(IgnoreApi = true)]
+    [Route("/error")]
+    public IActionResult HandleError() => Problem();
 }
