+++ src/server/HelloCodeaholics.Api/Program.cs
@@ -1,5 +1,4 @@
 using HelloCodeaholics.Data;
-using HelloCodeaholics.Infrastructure;
 using HelloCodeaholics.Services.Application;
 using HelloCodeaholics.Services.Interfaces;
