+++ src/server/HelloCodeaholics.Data/DataRepository.cs
@@ -1,11 +1,5 @@
-﻿using HelloCodeaholics.Infrastructure;
-using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
 using System.Data;
 
 namespace HelloCodeaholics.Data;
