+++ src/server/HelloCodeaholics.Data/Entities/Location.cs
@@ -1,0 +1,9 @@
+﻿namespace HelloCodeaholics.Data.Entities;
+
+public class Location
+{
+    public string Address { get; set; } = string.Empty;
+    public string City { get; set; } = string.Empty;
+    public string State { get; set; } = string.Empty;
+    public int Zip { get; set; }
+}
