+++ src/server/HelloCodeaholics.Data/HelloCodeDbContext.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using HelloCodeaholics.Data.Entities;
+﻿using HelloCodeaholics.Data.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace HelloCodeaholics.Infrastructure;
+namespace HelloCodeaholics.Data;
 
 public partial class HelloCodeDbContext : DbContext
 {
