+++ src/server/HelloCodeaholics.Services/Mapper/PharmacyMapper.cs
@@ -1,10 +1,5 @@
 ﻿using HelloCodeaholics.Common.Models;
 using HelloCodeaholics.Data.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace HelloCodeaholics.Services.Mapper;
